Projects/Vigil/requirements.txt
PyYAML>=6.0
httpx>=0.28
pywin32>=311
watchdog>=4.0
aiohttp>=3.9
